+++ src/mplfinance/_arg_validators.py
@@ -140,6 +140,11 @@ def _is_datelike(value):
             return False
     return False
 
+def _xlim_validator(value):
+    return (isinstance(value, (list,tuple)) and len(value) == 2
+            and (all([isinstance(v,(int,float)) for v in value])
+                 or all([_is_datelike(v) for v in value])))
+
 def _vlines_validator(value):
     '''Validate `vlines` kwarg value:  must be "datelike" or sequence of "datelike"
     '''